requirements.txt
numpy~=1.24.3
torch~=2.0.1
pybind11~=2.10.4
pyyaml~=6.0
